requirements.txt
aiogram==3.13.0
google-generativeai
python-dotenv
apscheduler